refactor(identity): domain model 패키지 분리 #48
