+++ systems/identity/identity-domain/src/test/kotlin/hs/kr/entrydsm/identity/domain/AccountTest.kt
@@ -1,6 +1,11 @@
 package hs.kr.entrydsm.identity.domain
 
+import hs.kr.entrydsm.identity.domain.enum.AccountStatus
+import hs.kr.entrydsm.identity.domain.enum.ApplicantStatus
+import hs.kr.entrydsm.identity.domain.enum.SignupType
 import hs.kr.entrydsm.identity.domain.exception.IdentityDomainException
+import hs.kr.entrydsm.identity.domain.model.Account
+import hs.kr.entrydsm.identity.domain.model.StudentProfile
 import java.time.Instant
 import java.time.LocalDate
 import org.junit.Assert.assertEquals
